+++ src/storage/interfaces/IBlockStorage.ts
