+++ src/storage/interfaces/IExitChannelStorage.ts
