+++ src/storage/interfaces/IStateSnapshotStorage.ts
